Refactor out IDB logic for testing
